Update staging models

+++ models/staging/stg_bike_trips.sql
@@ -13,7 +13,7 @@ SELECT
     CAST(end_date AS date) AS end_date,			
     CAST(end_time AS time) AS end_time,
     CAST(end_station_code AS int) AS end_station_code,			
-    CAST(duration_sec AS float) AS duration_sec,			
+    CAST(duration_sec AS FLOAT64) AS duration_sec,			
     CAST(is_member AS int) AS is_member,			
     CAST(yearid AS int) AS yearid			
 FROM BikeTrips
